requirements.txt
Django==5.1.3
django-admin-interface==0.29.2
django-extensions==3.2.3
django-imagekit==5.0.0
django-debug-toolbar==4.4.6
pilkit==3.0
pillow==11.0.0
python-magic==0.4.27
Werkzeug==3.1.3
django-environ==0.11.2
gunicorn==23.0.0
uvicorn==0.32.1
tzdata==2024.2
flake8==7.1.1
black==24.10.0